+++ src/main/java/com/estepper/estepper/service/FaseValoracionService.java
@@ -17,7 +17,7 @@ public interface FaseValoracionService {
     public List<FaseValoracion> faseValoracion(Participante participante);
     public void crearFormularios(Participante participante);
     public void crearFormulariosNuevos(Participante participante);
-    public void updateExploracion(String primeravez, Sexo sexo, Integer peso, Integer talla, Integer cmcintura, Integer edad, Integer imc, Participante participante);
+    public void updateExploracion(String primeravez, Sexo sexo, Double peso, Integer talla, Double cmcintura, Integer edad, Integer imc, Participante participante);
     public void updateFindrisc(Participante participante,Integer puntosedad, Integer puntosimc, Integer puntoscmcintura, Integer ptosactfisica, Integer ptosfrecfruta, Integer ptosmedicacion, Integer ptosglucosa, Integer ptosdiabetes, Integer puntuacion, String escalarriesgo);
     public void updateClasificacion(Clasificacion clasificacion, Participante participante);
     public void updateAntecedentes(Antecedentes antecedentes, Participante participante);
